scomposizione componenti
